data-and-ai-jhb/oct/code/claude-3/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
azure-ai-projects==1.0.0
azure-identity==1.25.1
pydantic==2.10.6
